agent-system-project/pyproject.toml
[project]
name = "agent-system-project"
version = "0.1.0"
description = ""
authors = [
    {name = "ahmed akram amer",email = "ahmedakram3ai@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "langgraph (>=1.1.10,<2.0.0)",
    "crewai (>=1.14.4,<2.0.0)",
    "fastapi (>=0.136.1,<0.137.0)",
    "uvicorn (>=0.46.0,<0.47.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "crewai-tools (>=1.14.4,<2.0.0)",
    "langchain-groq (>=1.1.2,<2.0.0)",
    "litellm (>=1.83.14,<2.0.0)"
]

[tool.poetry]
packages = [{include = "agent_system_project", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
